Change parameter of some contants and variables to uint64_t

+++ libraries/protocol/include/graphene/protocol/account.hpp
@@ -84,7 +84,7 @@ namespace graphene { namespace protocol {
       {
          uint64_t basic_fee      = 5*GRAPHENE_BLOCKCHAIN_PRECISION; ///< the cost to register the cheapest non-free account
          uint64_t premium_fee    = 2000*GRAPHENE_BLOCKCHAIN_PRECISION; ///< the cost to register the cheapest non-free account
-         uint32_t price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
+         uint64_t price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
       };
 
       asset           fee;
@@ -136,7 +136,7 @@ namespace graphene { namespace protocol {
       struct fee_parameters_type
       {
          share_type fee             = 20 * GRAPHENE_BLOCKCHAIN_PRECISION;
-         uint32_t   price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
+         uint64_t   price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
       };
 
       asset fee;
